Fix compatibility with TypeScript 4.7.x
TypeScript 4.7.x complains about the possible implicit conversion from
`symbol` to `string`. In order to resolve the issue we do not consider
known keys that are symbols as a possibility. Vuex cannot use them since
they cannot be serialized.

To reproduce the issue, you can upgrade the TypeScript version set as a
dev dependency and then run the test suite. You will see error like this
without this contribution:

```
 FAIL  tests/global-state.test.ts
  ● Test suite failed to run
    src/util.ts:52:48 - error TS2731: Implicit conversion of a 'symbol' to a 'string' will fail at runtime. Consider wrapping this expression in 'String(...)'.
    52   return cb(store, namespace ? `${namespace}/${prop}` : prop);
```